requirements.txt
openai
langchain-openai
langgraph
langchain-core
pyperclip
keyboard
win10toast
requests
langchain
pystray
pillow
pydantic
cryptography
cx-Freeze